# backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Worldwide Rapid Task For Money - Verification Automation Engine
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /verifications:
    get:
      operationId: listVerifications
      tags: [verifications]
      summary: List verification results
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, auto_approved, auto_rejected, manual_review, approved, rejected]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of verifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationList"
  /verifications/{id}:
    get:
      operationId: getVerification
      tags: [verifications]
      summary: Get a verification result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"
  /verifications/{id}/review:
    post:
      operationId: reviewVerification
      tags: [verifications]
      summary: Manual review of a verification (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "200":
          description: Reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"

  # Users (admin)
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [client, worker, admin]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/reputation:
    get:
      operationId: getUserReputation
      tags: [users]
      summary: Get a user's reputation details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reputation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReputationDetail"

  # Wallet
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [client, worker, admin]
        country:
          type: ["string", "null"]
        language:
          type: string
          enum: [fr, en]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        reputationScore:
          type: number
        createdAt:
          type: string
          format: date-time
    UserList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        country:
          type: string
        language:
          type: string
          enum: [fr, en]
        avatarUrl:
          type: string
        isActive:
          type: boolean
    ReputationDetail:
      type: object
      required: [userId, score, tasksCompleted, validationRate, avgProofQuality, level, badges]
      properties:
        userId:
          type: integer
        score:
          type: number
        tasksCompleted:
          type: integer
        validationRate:
          type: number
        avgProofQuality:
          type: number
        level:
          type: string
          enum: [newcomer, bronze, silver, gold, platinum]
        badges:
          type: array
          items:
            type: string
        topCategories:
          type: array
          items:
            type: string

    # Campaign
...
    VerificationResult:
      type: object
      required: [id, assignmentId, status, confidenceScore, method, createdAt]
      properties:
        id:
          type: integer
        assignmentId:
          type: integer
        status:
          type: string
          enum: [pending, auto_approved, auto_rejected, manual_review, approved, rejected]
        confidenceScore:
          type: number
        method:
          type: string
          enum: [automatic, manual, hybrid]
        checks:
          type: array
          items:
            $ref: "#/components/schemas/VerificationCheck"
        reviewNotes:
          type: ["string", "null"]
        reviewedBy:
          type: ["integer", "null"]
        reviewedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    VerificationCheck:
      type: object
      required: [name, passed, score]
      properties:
        name:
          type: string
        passed:
          type: boolean
        score:
          type: number
        details:
          type: ["string", "null"]
    VerificationList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/VerificationResult"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ReviewInput:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [approved, rejected]
        notes:
          type: string

    # Wallet
...