# backend/src/api-zod/generated/types/reputationDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Worldwide Rapid Task For Money - Verification Automation Engine
 * OpenAPI spec version: 0.1.0
 */
import type { ReputationDetailLevel } from './reputationDetailLevel';

export interface ReputationDetail {
  userId: number;
  score: number;
  tasksCompleted: number;
  validationRate: number;
  avgProofQuality: number;
  level: ReputationDetailLevel;
  badges: string[];
  topCategories?: string[];
}