# backend/src/api-zod/generated/types/reviewInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Worldwide Rapid Task For Money - Verification Automation Engine
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewInputDecision } from './reviewInputDecision';

export interface ReviewInput {
  decision: ReviewInputDecision;
  notes?: string;
}